% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.model.R
\docType{class}
\name{Combinatorics-class}
\alias{Combinatorics-class}
\alias{initialize,Combinatorics-method}
\alias{combinatorics.model}
\title{Combinatorics model}
\usage{
\S4method{initialize}{Combinatorics}(.Object, ..., data = matrix(),
  arrayLogicalConnections = character())

combinatorics.model(...)
}
\arguments{
\item{.Object}{Комбинаторная модель}

\item{...}{аргументы конструктура}

\item{data}{объект класса \code{"matrix"}, содержащий матрицу, где в строках указаны
мероприятия, а в столбцах показатели мероприятий (ресурсные ограничение и целевые
показатели).}

\item{arrayLogicalConnections}{объект класса \code{"vector"}, содержащий
перечень логических связей в виде логических формул.}
}
\value{
Комбинаторная модель
}
\description{
Комбинаторные модели - это модели поиска оптимальных решений при наличии ресурсных
ограничений, логических связей и целей. Примерами моделей в частности является:
\itemize{
\item Компоновка инженерно-экономического проекта;
\item Построение системы, оптимальной по надежности;
\item Управление работами;
\item Управление портфелем ценных бумаг;
\item Предупреждение чрезвычайных ситуаций.
}

Используйте метод combinatorics.model для создания новой комбинаторной модели.
}
\note{
Во всех случаях, когда имеются варианты мероприятий, ресурсные ограничения, логические
связи и цели возникает комбинаторная задача. Практика свидетельствует о наличии соотношения
1:10:100:1000, где один доллар - «экономия» на принятии упрощенного решения на стадии
его формирования, а 10, 100, 1000 - потери на последующих стадиях жизненного цикла решения.

\strong{Мероприятие} - составная часть инвестиционного,  технического, правового или  иного проекта,
требующая затрат,  приносящая результаты и увязанная с другими мероприятиями логическими
связями. Мероприятие, включенное в логическую сеть, называется  элементом. Мероприятие так же
означает элемент множества. Решение задачи булевого программирования представляет собой
набор мероприятий (элементов), не противоречащий сети логических связей, требующий затрат
ограниченного ресурса в заданном пределе и доставляющий максимум (или минимум) целевому
показателю.

\strong{Показатели} - характеристика мероприятия в комбинаторной модели, имеющая количественное
выражение. Предполагается, что показатели аддитивны, то есть, что для любого набора мероприятий
соответствующие показатели можно складывать. Среди показателей один или несколько
выбирается в качестве ограниченного ресурса и один в качестве целевого показателя.
Ограниченный ресурс - являющется дефицитным или  лимитированным.
Целевые показатели - принимаемые в качестве критерия при дискретной оптимизации.

\strong{Логическая связь} (ЛС) - совместимость, обязательная последовательность осуществления или
обязательная необходимость  осуществления мероприятий, вытекающая из технологических,
технических, организационных или правовых условий их осуществления (в математической логике
используется термин "логический оператор"). Логическя связь записываются в виде логической формулы.

\strong{Логическая формула} - логическая связь между мероприятий, выраженная одним из следующих способов:
\itemize{
\item \code{[…,…,…]} – в квадратных скобках перечислены номера мероприятий, разделенные запятой, которые
исключают друг друга (логическая связь «или»);
\item \code{(…,…,…)->(…,…,…)}, - в скобках, слева от стрелки, дан перечень номеров мероприятий,
разделенных запятой, присутствие каждого из которых в решении является обязательным условием
для того, чтобы решение могло содержать все или некоторые из мероприятий, номера которых
перечислены в скобках, справа от стрелки (логическая связь «если…то…»);
\item \code{{…,…,…}} – в фигурных скобках перечислены номера мероприятий, разделенные запятой, среди
которых хотя бы одно непременно должно быть включено в решение
(логическая связь «должен присутствовать хотя бы один из…»;
\item \code{…} - номер мероприятия в отдельной строке означает мероприятие, логически независимое от
остальных («Независимое мероприятие»)
}

\strong{Логическая сеть} - совокупность мероприятий  с наложенными на них логическими связями.
Логическая сеть вводится и редактируется как совокупность логических связей, включающих,
в том числе, независимые мероприятия (иначе  говоря,  мероприятия,  не имеющие логических
связей с другими мероприятиями).
}
\section{Slots}{

\describe{
\item{\code{parameters}:}{объект класса \code{"list"}, содержащий показатели модели}
\item{\code{arrangements}:}{объект класса  \code{"list"}, содержащий варианты мероприятий модели}
\item{\code{logicconnections}:}{объект класса \code{"list"}, содержащий логическую сеть модели}
\item{\code{name}:}{объект класса \code{"character"}, содержащий наименование модели}
}
}

\examples{
  model <- combinatorics.model(name="Example")
}
