% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.add.R
\docType{methods}
\name{combinatorics.add.arrangement}
\alias{combinatorics.add.arrangement}
\alias{combinatorics.add.arrangement,Combinatorics,character,vector-method}
\title{Method add arrangement}
\usage{
combinatorics.add.arrangement(model, name, values)

\S4method{combinatorics.add.arrangement}{Combinatorics,character,vector}(model,
  name, values)
}
\arguments{
\item{model}{Combinatorics model}

\item{name}{name model}

\item{values}{numeric vector}
}
\value{
Combinatroics model
}
\description{
Add arrangement in model.
}
\examples{

  model <- combinatorics.example.Model3()
  combinatorics.add.arrangement(model, 'M25', c(10.1, 20.2, 30.3, 40.4, 50.5, 60.6))
}
