% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.add.R
\docType{methods}
\name{combinatorics.add.logicconnection}
\alias{combinatorics.add.logicconnection}
\alias{combinatorics.add.logicconnection,Combinatorics,character-method}
\title{Method add logicconnection in model.}
\usage{
combinatorics.add.logicconnection(model, logicconnection)

\S4method{combinatorics.add.logicconnection}{Combinatorics,character}(model,
  logicconnection)
}
\arguments{
\item{model}{Combinatorics model}

\item{logicconnection}{logicconnection value}
}
\value{
Combinatroics model
}
\description{
Add logicconnection in model.
}
\examples{

  model <- combinatorics.example.Model3()
  combinatorics.add.logicconnection(model, "(1)->(7,18,22,24)")
}
