% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.add.R
\docType{methods}
\name{combinatorics.add.parameter}
\alias{combinatorics.add.parameter}
\alias{combinatorics.add.parameter,Combinatorics,character,character,numeric,missing-method}
\alias{combinatorics.add.parameter,Combinatorics,character,missing,numeric,missing-method}
\alias{combinatorics.add.parameter,Combinatorics,character,character,missing,missing-method}
\alias{combinatorics.add.parameter,Combinatorics,character,missing,missing,missing-method}
\alias{combinatorics.add.parameter,Combinatorics,character,character,numeric,vector-method}
\alias{combinatorics.add.parameter,Combinatorics,character,character,missing,vector-method}
\alias{combinatorics.add.parameter,Combinatorics,character,missing,numeric,vector-method}
\title{Add parameter in model}
\usage{
combinatorics.add.parameter(model, name, dimension, value, arrangementValues)


  \S4method{combinatorics.add.parameter}{Combinatorics,character,character,numeric,missing}(model,
  name, dimension, value, arrangementValues)


  \S4method{combinatorics.add.parameter}{Combinatorics,character,missing,numeric,missing}(model,
  name, dimension, value, arrangementValues)


  \S4method{combinatorics.add.parameter}{Combinatorics,character,character,missing,missing}(model,
  name, dimension, value, arrangementValues)


  \S4method{combinatorics.add.parameter}{Combinatorics,character,missing,missing,missing}(model,
  name, dimension, value, arrangementValues)


  \S4method{combinatorics.add.parameter}{Combinatorics,character,character,numeric,vector}(model,
  name, dimension, value, arrangementValues)


  \S4method{combinatorics.add.parameter}{Combinatorics,character,character,missing,vector}(model,
  name, dimension, value, arrangementValues)


  \S4method{combinatorics.add.parameter}{Combinatorics,character,missing,numeric,vector}(model,
  name, dimension, value, arrangementValues)
}
\arguments{
\item{model}{объект класса \code{"Combinatorics"}, содержащий комбинаторную модель}

\item{name}{объект класса \code{"character"}, содержащий наименование добавляемого
показателя}

\item{dimension}{объект класса \code{"character"}, содержащий размернойть добавляемого
показателя}

\item{value}{объект класса \code{"numeric"}, содержащий значение огранченного ресурса.}

\item{arrangementValues}{объект класса \code{"vector"}, содержащий числовые значения показателя для.
всех мероприятий в комбинаторной модели. Количество значений в векторе должно быть равно количеству
мероприятий в комбинаторной модели}
}
\value{
возвращает объект класса \code{"Combinatorics"} с внесёнными изменениями
}
\description{
Добавляет показатель в комбинаторную модель. При указании значения в парамете \code{"value"}
  показатель используется  в качестве как ограниченного ресурса
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,name = character,dimension = character,value = numeric,arrangementValues = missing}: добавление ограниченного ресурса в комбинаторную модель
не содержащей мероприятий

\item \code{model = Combinatorics,name = character,dimension = missing,value = numeric,arrangementValues = missing}: добавление ограниченного ресурса без указания его размерности
в комбинаторную модель не содержащей мероприятий

\item \code{model = Combinatorics,name = character,dimension = character,value = missing,arrangementValues = missing}: добавление показателя в комбинаторную модель не
содержащей мероприятий

\item \code{model = Combinatorics,name = character,dimension = missing,value = missing,arrangementValues = missing}: добавление показателя без указания его размерности
в комбинаторную модель не содержащей мероприятий

\item \code{model = Combinatorics,name = character,dimension = character,value = numeric,arrangementValues = vector}: добавление ограниченного ресурса в комбинаторную модель

\item \code{model = Combinatorics,name = character,dimension = character,value = missing,arrangementValues = vector}: добавление показателя в комбинаторную модель

\item \code{model = Combinatorics,name = character,dimension = missing,value = numeric,arrangementValues = vector}: добавление ограниченного ресурса без указания его размерности
в комбинаторную модель
}}

\examples{

  model <- combinatorics.model()
  combinatorics.add.parameter (model, "R1", "pc", 10)

  model <- combinatorics.model()
  combinatorics.add.parameter (model, "R1", value=10)

  model <- combinatorics.model()
  combinatorics.add.parameter (model, "R1", "pc.")

  model <- combinatorics.model()
  combinatorics.add.parameter (model, "R1")

  model <- combinatorics.example.Model2()
  combinatorics.add.parameter (model, "Electricity", "Thousand kWh/hour", 75,
     c(39, 24.3, 17.7, 26.7, 11.1, 10.8, 18.3, 12.3))

  model <- combinatorics.example.Model2()
  combinatorics.add.parameter (model, "Electricity", "Thousand kWh/hour",
     arrangementValues=c(39, 24.3, 17.7, 26.7, 11.1, 10.8, 18.3, 12.3))

  model <- combinatorics.example.Model2()
  combinatorics.add.parameter (model, "Electricity", value=75,
     arrangementValues=c(39, 24.3, 17.7, 26.7, 11.1, 10.8, 18.3, 12.3))
}
