% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.calculate.R
\name{combinatorics.calculate}
\alias{combinatorics.calculate}
\title{Calculation of the combinatorics model}
\usage{
combinatorics.calculate(model, maximum = TRUE, ...)
}
\arguments{
\item{model}{объект класса \code{"Combinatorics"}, содержащий комбинаторную модель задачи}

\item{maximum}{объект класса \code{"boolean"}, содержащий флаг указания расчёта на максими или
минимум.}

\item{...}{дополнительные аргументы, передаваемые другим методам или получаемые от них.}
}
\value{
объект, содержаший результаты расчёта комбинаторной модели на максимум
 или минимум
}
\description{
Расчёт комбинаторной модели на максимум, т.е. максимизируется значение целевого показателя - «притока»
или на минимум, т.е. предел допустимого совокупного «притока». Смысл расчёта на миниму в том, что
Установив заранее желаемую цель, можной найти те минимально необходимые затраты, которые
обеспечат достижение этой цели.

Для выполнения расчёта должны быть указаные значения пределов ограниченных ресурсов и выбран
целевой показатель.
}
\examples{
  combinatorics.calculate(combinatorics.example.Model1())

  combinatorics.calculate(combinatorics.example.Model2())

  result <- combinatorics.calculate(combinatorics.example.Model1(), maximum=FALSE)
  print(result, nameArrangment=TRUE)
  plot(result, type="o")
}
