% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.models.R
\name{Model1}
\alias{Model1}
\alias{combinatorics.example.Model1}
\title{Combinatorial model of the simplest problem}
\usage{
combinatorics.example.Model1()
}
\value{
возвращает объект класса \code{"Combinatorics"} с исходными данными задачи
}
\description{
Пример простейшей комбинаторной модели, для задачи в которой фирме предполагает купить станок.
При этом возникают варианты мероприятий, представленные в следующей таблице:
\tabular{cllrr}{
№ \tab Название мероприятия \tab Описание мероприятия \tab Затраты, тыс.руб \tab Прибыль в год, тыс.руб. \cr
1 \tab Станок - А       \tab Покупка станка марки «А» \tab 350 \tab 117 \cr
2 \tab Станок - Б       \tab Покупка станка марки «Б» \tab 420 \tab 125 \cr
3 \tab Аренда - А       \tab Аренда станка марки «А»  \tab 330 \tab 105 \cr
4 \tab Оборудование к А \tab Покупка дополнительного оборудования к станку марки «А» \tab 77 \tab 83 \cr
5 \tab Оборудование к Б \tab Покупка дополнительного оборудования к станку марки «Б» \tab 75 \tab 81
}

Из приведенных вариантов мероприятий проистекает следующий набор логических связей между мероприятиями:
\itemize{
\item \code{[1,2,3]}  - мероприятие № 1, мероприятие № 2, мероприятие № 3  взаимно друг друга исключают
\item \code{{1,2,3}} - какое-либо из мероприятий № 1, № 2, или № 3 обязательно должно быть исполнено;
\item \code{(1)->(4), (3)->(4)} - если будет исполнено мероприятие № 1 или мероприятие № 3 то может быть исполнено и мероприятие № 4;
\item \code{(2)->(5)} - если будет исполнено мероприятие № 2 то может быть исполнено и мероприятие № 5.
}

Ограниченным ресурсом являются средства на покупку станка и оборудования. Значение ограниченного
ресурса задано в 700 т.руб. Целевым показателем является получение максимальной прибыли.
}
\examples{
  model1 <- combinatorics.example.Model1()
  combinatorics.calculate(model1)
  print(combinatorics.calculate(model1), nameArrangment = TRUE)

}
