% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.models.R
\name{Model2}
\alias{Model2}
\alias{combinatorics.example.Model2}
\title{Combinatorics model of the simple optimal integrated investment program}
\usage{
combinatorics.example.Model2()
}
\description{
Пример комбинаторной модели, для задачи формирования оптимальной комплексной
инвестиционной программы предприятия.
}
\note{
«Даунбрукс» - агропромышленное предприятие, расположенное в
Московской области недалеко от аэропорта. Предприятие специализируется на
производстве грибов-шампиньонов и имеет незавершенное строительство в виде
крупного павильона, в котором могут быть размещены различные производства.
Кроме того, оно имеет некоторые свободные земельные участки и располагает
ограниченными собственными средствами для финансирования проектов.
Однако предприятие может привлечь соинвесторов на взаимно выгодных
условиях.

С помощью специалистов предприятия и соинвесторов было разработано
множество возможных вариантов инвестиционных проектов (ИП), которые в
совокупности и образовали портфель ИП. Для каждого ИП рассчитан валовой
дисконтированный на весь горизонт проекта доход («приток») и совокупные на
весь горизонт проекта дисконтированные затраты («отток»)
\tabular{clrr}{
№ \tab Название инвестиционного проекта \tab Отток, млн.руб. \tab Приток, млн.руб. \cr
1 \tab Расширение шампиньонного комплекса \tab 5.40 \tab 7.60 \cr
2 \tab Производство упаковки \tab 3.70 \tab 4.40 \cr
3 \tab Таможенный склад \tab 2.70 \tab 3.20 \cr
4 \tab Склад продукции \tab 3.30 \tab 5.60 \cr
5 \tab Производство компоста \tab 1.00 \tab 1.70 \cr
6 \tab Производство рыбы \tab 1.30 \tab 2.30 \cr
7 \tab Теплица 1 \tab 2.50 \tab 3.60 \cr
8 \tab Теплица 2 \tab 1.60 \tab 2.50
}

Предполагается, что между заданным множеством инвестиционных
проектов имеются некоторые логические связи (ЛС), которые могут быть
представлены в следующей содержательной форме:
\itemize{
\item \emph{«Расширение шампиньонного комплекса»} - предполагается расширить
производственные площади и модернизировать действующее производство.
\item \emph{«Производство упаковки»} – в настоящее время упаковка для грибов
закупается. Данный инвестиционный проект предполагает производить
упаковку собственными силами. При этом производство предполагается
расположить в недостроенном помещении.
\item \emph{«Таможенный склад»} – поскольку шампиньонный комплекс
расположен недалеко от аэропорта, то имеющееся недостроенное помещение
может быть достроено, оборудовано соответствующим образом и сдано в
аренду аэропорту.
\item \emph{«Склад продукции»} - реконструкция и переоборудование действующего
склада.
\item \emph{«Производство компоста»} - компост для производства грибов
производится на предприятии в недостаточном количестве, а остальное
закупается. Данный проект позволит настолько увеличить производство
компоста, что его будет достаточно как для собственных нужд, так и для
продажи.
\item \emph{«Производство рыбы»} – строительство бассейна для промышленного
выращивания рыбы.
\item \emph{«Теплица 1»} – строительство теплицы для выращивания овощей с
независимым водоснабжением.
\item \emph{«Теплица 2»} – строительство теплицы на тех же площадях, что и в
варианте 1, но с водоснабжением из бассейна для рыбы, а так же с
получением части удобрения из бассейна.
}

Общая сумма средств, которое предприятие может использовать для
инвестиций на все инвестиционные проекты, составляет \strong{11 млн. руб.}
Задача состоит в том, чтобы, не выходя за заданный лимит финансовых
средств и не нарушая ни одного из логических условий сформировать
портфель инвестиционных проектов, у которого суммарный «приток» будет
максимален. Этот портфель ИП и будет оптимальной комплексной
инвестиционной программой предприятия.
}
\examples{
  model2 <- combinatorics.example.Model2()
  combinatorics.calculate(model2)

}
