% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics-package.r
\docType{package}
\name{combinatorics}
\alias{combinatorics}
\alias{combinatorics-package}
\title{An implementation algorithm of discrete optimization}
\description{
The solution of combinatorial problems by the dynamic programming
algorithm for solving Boolean problems. The algorithm provides the
search for optimal solutions in the presence of resource constraints,
 logical relationships and goals.
}
