% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.remove.R
\docType{methods}
\name{combinatorics.remove.arrangement}
\alias{combinatorics.remove.arrangement}
\alias{combinatorics.remove.arrangement,Combinatorics,character-method}
\alias{combinatorics.remove.arrangement,Combinatorics,numeric-method}
\title{Remove arrangement}
\usage{
combinatorics.remove.arrangement(model, arrangement)

\S4method{combinatorics.remove.arrangement}{Combinatorics,character}(model,
  arrangement)

\S4method{combinatorics.remove.arrangement}{Combinatorics,numeric}(model,
  arrangement)
}
\arguments{
\item{model}{объект класса \code{"Combinatorics"}, содержащий комбинаторную модель}

\item{arrangement}{объект класса \code{"character"} или \code{"numeric"}, содержащий
наименование мероприятия или индекс мероприятия в списке мероприятий}
}
\value{
возвращает объект класса \code{"Combinatorics"} с внесёнными изменениями
}
\description{
Исключает мероприятие из комбинаторной модели
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,arrangement = character}: поиск мероприятия по наименованию и исключение его

\item \code{model = Combinatorics,arrangement = numeric}: поиск мероприятия по индексe в списке мероприятий
и исключение его
}}

\examples{

  model <- combinatorics.example.Model3()
  combinatorics.remove.arrangement(model, 'M1')

  model <- combinatorics.example.Model3()
  combinatorics.remove.arrangement(model, 1)
}
\seealso{
\code{\link{combinatorics.remove.logicconnection}},
\code{\link{combinatorics.remove.parameter}}
}
