% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.remove.R
\docType{methods}
\name{combinatorics.remove.logicconnection}
\alias{combinatorics.remove.logicconnection}
\alias{combinatorics.remove.logicconnection,Combinatorics,character-method}
\alias{combinatorics.remove.logicconnection,Combinatorics,numeric-method}
\title{Remove logicconnection}
\usage{
combinatorics.remove.logicconnection(model, logicconnection)

\S4method{combinatorics.remove.logicconnection}{Combinatorics,character}(model,
  logicconnection)

\S4method{combinatorics.remove.logicconnection}{Combinatorics,numeric}(model,
  logicconnection)
}
\arguments{
\item{model}{Combinatorics model}

\item{logicconnection}{logicconnection}
}
\value{
Combinatroics model
}
\description{
Remove logicconnection from model
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,logicconnection = character}: remove by logic value

\item \code{model = Combinatorics,logicconnection = numeric}: remove by index
}}

\examples{

  model <- combinatorics.example.Model3()
  combinatorics.remove.logicconnection(model, "[1,4,5]")

  model <- combinatorics.example.Model3()
  combinatorics.remove.logicconnection(model, 4)
}
