% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.remove.R
\docType{methods}
\name{combinatorics.remove.parameter}
\alias{combinatorics.remove.parameter}
\alias{combinatorics.remove.parameter,Combinatorics,character-method}
\alias{combinatorics.remove.parameter,Combinatorics,numeric-method}
\title{Remove parameter}
\usage{
combinatorics.remove.parameter(model, parameter)

\S4method{combinatorics.remove.parameter}{Combinatorics,character}(model,
  parameter)

\S4method{combinatorics.remove.parameter}{Combinatorics,numeric}(model,
  parameter)
}
\arguments{
\item{model}{Combinatorics model}

\item{parameter}{parameter}
}
\value{
Combinatroics model
}
\description{
Remove parameter from model
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,parameter = character}: remove by name

\item \code{model = Combinatorics,parameter = numeric}: remove by index
}}

\examples{

  model <- combinatorics.example.Model3()
  combinatorics.remove.parameter(model, "R1")

  model <- combinatorics.example.Model3()
  combinatorics.remove.parameter(model, 4)
}
