% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.set.R
\docType{methods}
\name{combinatorics.set.arrangement}
\alias{combinatorics.set.arrangement}
\alias{combinatorics.set.arrangement,Combinatorics,character,vector-method}
\alias{combinatorics.set.arrangement,Combinatorics,numeric,vector-method}
\title{Set arrangement values}
\usage{
combinatorics.set.arrangement(model, arrangement, values)

\S4method{combinatorics.set.arrangement}{Combinatorics,character,vector}(model,
  arrangement, values)

\S4method{combinatorics.set.arrangement}{Combinatorics,numeric,vector}(model,
  arrangement, values)
}
\arguments{
\item{model}{объект класса \code{"Combinatorics"}, содержащий комбинаторную модель}

\item{arrangement}{объект класса \code{"character"} или \code{"numeric"}, содержащий
наименование мероприятия или индекс мероприятия в списке мероприятий}

\item{values}{объект класса \code{"vector"}, содержащий числовые значения показателей.
Количество значений в векторе должно быть равно колчеству показателей в комбинаторной модели}
}
\value{
возвращает объект класса \code{"Combinatorics"} с внесёнными изменениями
}
\description{
Присваивает одному из мероприятий в комбинаторной модели значения показателей.
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,arrangement = character,values = vector}: поиск мероприятия по наименованию и установка новых
значений показателей.

\item \code{model = Combinatorics,arrangement = numeric,values = vector}: поиск мероприятия по индексe в списке мероприятий
и установка новых значений показателей.
}}

\examples{

  model <- combinatorics.example.Model3()
  combinatorics.set.arrangement(model, "M0", c(10, 20, 30, 40, 50, 60.6))

  model <- combinatorics.example.Model3()
  combinatorics.set.arrangement(model, 1, c(10.1, 20.2, 30.3, 40.4, 50.5, 60.6))
}
