% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.set.R
\docType{methods}
\name{combinatorics.set.name}
\alias{combinatorics.set.name}
\alias{combinatorics.set.name,Combinatorics,character-method}
\title{Method set model name}
\usage{
combinatorics.set.name(model, name)

\S4method{combinatorics.set.name}{Combinatorics,character}(model, name)
}
\arguments{
\item{model}{Combinatorics model}

\item{name}{model name}
}
\value{
Combinatroics model
}
\description{
Set model name
}
\examples{

  model <- combinatorics.example.Model3()
  combinatorics.set.name (model, 'Demo Name')
}
