% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.set.R
\docType{methods}
\name{combinatorics.set.target}
\alias{combinatorics.set.target}
\alias{combinatorics.set.target,Combinatorics,character,logical-method}
\alias{combinatorics.set.target,Combinatorics,numeric,logical-method}
\title{Set target in model}
\usage{
combinatorics.set.target(model, parameter, value)

\S4method{combinatorics.set.target}{Combinatorics,character,logical}(model,
  parameter, value)

\S4method{combinatorics.set.target}{Combinatorics,numeric,logical}(model,
  parameter, value)
}
\arguments{
\item{model}{объект класса \code{"Combinatorics"}, содержащий комбинаторную модель}

\item{parameter}{объект класса \code{"character"} или \code{"numeric"}, содержащий
наименование показателя или индекс показателя в списке}

\item{value}{объект класса \code{"boolean"}, содержащий флаг использования показателя
как целевого показателя. TRUE - показатель используется как целевой показатель.}
}
\value{
возвращает объект класса \code{"Combinatorics"} с внесёнными изменениями
}
\description{
Присваивает одному из показателей комбинаторной модели флаг использования его в качестве
  целевого показателя или снимает снимает флаг. В комбинаторной модели при расчёте может быть
  только один целевой показатель.
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,parameter = character,value = logical}: поиск показателя по наименованию и установка флага
его использования в качестве целевого показателя.

\item \code{model = Combinatorics,parameter = numeric,value = logical}: поиск показателя по индексу в списке показателей и установка флага
его использования в качестве целевого показателя.
}}

\examples{

  model <- combinatorics.example.Model3()
  combinatorics.set.target (model, 'R1', TRUE)

  model <- combinatorics.example.Model3()
  combinatorics.set.target (model, 4, TRUE)
}
