% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.plot.R
\name{plot.combinatorics.result}
\alias{plot.combinatorics.result}
\title{plot}
\usage{
\method{plot}{combinatorics.result}(x, limit = 1, TargetDivLimit = FALSE,
  ...)
}
\arguments{
\item{x}{Combinatorics model}

\item{limit}{index}

\item{TargetDivLimit}{y axis}

\item{...}{other arguments}
}
\description{
plot
}
\examples{
  out <- combinatorics.calculate(combinatorics.example.Model1())
  plot(out, type="o")
  plot(out, type="o", TargetDivLimit=TRUE)

}
