% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.print.R
\name{print.combinatorics.result}
\alias{print.combinatorics.result}
\title{print}
\usage{
\method{print}{combinatorics.result}(x, nameArrangment = FALSE,
  digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Combinatorics model}

\item{nameArrangment}{boolean}

\item{digits}{integer value}

\item{...}{other arguments}
}
\description{
print
}
\examples{
  out <- combinatorics.calculate(combinatorics.example.Model1())
  print(out)
  print(out, nameArrangment = TRUE)
}
