% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stirling2.R
\name{stirling2}
\alias{stirling2}
\alias{Stirling2}
\title{Prints the value of Stirling numbers of second kind}
\usage{
Stirling2(n, k)
}
\arguments{
\item{n}{the first parameter representing the number of elements in the set total.}

\item{k}{the second parameter representing the number of groups to be formed.}
}
\value{
Stirling2: the determined value of Stirling numbers of second kind.
}
\description{
Determining the Stirling number of second kind.
}
\details{
Stirling numbers of second kind is a very useful term used in combinatorics denoting the number of all possible groups of size k from a set of size n.
}
\examples{
Stirling2(3,2)
}
\references{
Bóna,Miklós(2017,ISBN 9789813148840).
}
\author{
Anik Paul
}
