% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.model.R
\docType{class}
\name{Combinatorics-class}
\alias{Combinatorics-class}
\alias{initialize,Combinatorics-method}
\alias{combinatorics.model}
\title{Комбинаторная модель}
\usage{
\S4method{initialize}{Combinatorics}(.Object, ..., data = matrix(),
  arrayLogicalConnections = character())

combinatorics.model(...)
}
\arguments{
\item{.Object}{Комбинаторная модель}

\item{...}{аргументы конструктура}

\item{data}{объект класса \code{"matrix"}, содержащий матрицу, где в строках
указаны мероприятия, а в столбцах ресурсные ограничение и цели.}

\item{arrayLogicalConnections}{объект класса \code{"vector"}, содержащий
перечень логических связей}
}
\value{
Комбинаторная модель
}
\description{
Комбинаторная модель

Используйте функцию combinatorics.model для создания новой комбинаторной модель
}
\section{Slots}{

 \describe{
   \item{\code{parameters}:}{объект класса \code{"list"}, содержащий ресурсные ограничения и цели.}
   \item{\code{arrangements}:}{объект класса  \code{"list"}, содержащий варианты мероприятий.}
   \item{\code{logicconnections}:}{объект класса \code{"list"}, содержащий логические связи мероприятий.}
   \item{\code{name}:}{объект класса \code{"character"}, содержащий наименование модели.}
 }
}

\examples{
  model <- combinatorics.model(name="Example")
}
