% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.model.R
\docType{class}
\name{Combinatorics-class}
\alias{Combinatorics-class}
\alias{Combinatorics}
\alias{combinatorics.model}
\title{Combinatorics Model}
\usage{
combinatorics.model(...)
}
\arguments{
\item{...}{constructor arguments}
}
\description{
Combinatorics Model

Use combinatorics.model for create new Combinatoric sobject
}
\section{Slots}{

 \describe{
   \item{\code{parameters}:}{Object of class \code{"list"}, containing data from of resources and targets}
   \item{\code{arrangements}:}{Object of class \code{"list"}, containing data from of arrangements.}
   \item{\code{logicconnections}:}{Object of class \code{"list"}, containing data logic network.}
   \item{\code{name}:}{Object of class \code{"character"}, containing data that needs to go in name.}
 }
}

\examples{
  model <- combinatorics.model(name="Example")
}
