% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.add.R
\docType{methods}
\name{combinatorics.add.arrangement}
\alias{combinatorics.add.arrangement}
\alias{combinatorics.add.arrangement,Combinatorics,character,vector-method}
\title{Method add arrangement}
\usage{
combinatorics.add.arrangement(model, name, values)

\S4method{combinatorics.add.arrangement}{Combinatorics,character,vector}(model,
  name, values)
}
\arguments{
\item{model}{объект класса \code{"Combinatorics"}, содержащий комбинаторную модель}

\item{name}{объект класса \code{"character"}, содержащий наименование добавляемого
мероприятия}

\item{values}{объект класса \code{"vector"}, содержащий содержащий числовые значения показателей. Количество
количество значений в векторе должно быть равно колчеству показателей в комбинаторной модели}
}
\value{
возвращает объект класса \code{"Combinatorics"} с внесёнными изменениями
}
\description{
Добавляет мероприятие в комбинаторную модель.
}
\examples{

  model <- combinatorics.example.Model3()
  combinatorics.add.arrangement(model, 'M25', c(10.1, 20.2, 30.3, 40.4, 50.5, 60.6))
}
