% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.calculate.R
\name{combinatorics.calculate}
\alias{combinatorics.calculate}
\title{combinatorics.calculate}
\usage{
combinatorics.calculate(model, maximum = TRUE, ...)
}
\arguments{
\item{model}{Combinatorics model}

\item{maximum}{boolean}

\item{...}{other arguments}
}
\value{
object
}
\description{
combinatorics.calculate
}
\examples{
  combinatorics.calculate(combinatorics.example.Model1())

  combinatorics.calculate(combinatorics.example.Model2())

  result <- combinatorics.calculate(combinatorics.example.Model1(), maximum=FALSE)
  print(result, nameArrangment=TRUE)
  plot(result, type="o")
}
