% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.models.R
\name{Model1}
\alias{Model1}
\alias{combinatorics.example.Model1}
\title{Model1}
\usage{
combinatorics.example.Model1()
}
\description{
Model1
}
\examples{
  model1 <- combinatorics.example.Model1()
  combinatorics.calculate(model1)
  print(combinatorics.calculate(model1), nameArrangment = TRUE)

}
