% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.models.R
\name{Model2}
\alias{Model2}
\alias{combinatorics.example.Model2}
\title{Model2}
\usage{
combinatorics.example.Model2()
}
\description{
Model2
}
\examples{
  model2 <- combinatorics.example.Model2()
  combinatorics.calculate(model2)

}
