% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.models.R
\name{Model3}
\alias{Model3}
\alias{combinatorics.example.Model3}
\title{Model3}
\usage{
combinatorics.example.Model3()
}
\description{
Model3
}
\examples{
  model3 <- combinatorics.example.Model3()
  combinatorics.calculate(model3)

}
