% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.remove.R
\docType{methods}
\name{combinatorics.remove.arrangement}
\alias{combinatorics.remove.arrangement}
\alias{combinatorics.remove.arrangement,Combinatorics,character-method}
\alias{combinatorics.remove.arrangement,Combinatorics,numeric-method}
\title{Remove arrangement}
\usage{
combinatorics.remove.arrangement(model, arrangement)

\S4method{combinatorics.remove.arrangement}{Combinatorics,character}(model,
  arrangement)

\S4method{combinatorics.remove.arrangement}{Combinatorics,numeric}(model,
  arrangement)
}
\arguments{
\item{model}{Combinatorics model}

\item{arrangement}{arrangement}
}
\value{
Combinatroics model
}
\description{
Remove arrangement from model
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,arrangement = character}: remove arrangement by name

\item \code{model = Combinatorics,arrangement = numeric}: remove by arrangement index
}}

\examples{

  model <- combinatorics.example.Model3()
  combinatorics.remove.arrangement(model, 'M1')

  model <- combinatorics.example.Model3()
  combinatorics.remove.arrangement(model, 1)
}
\seealso{
\code{\link{combinatorics.remove.logicconnection}},
\code{\link{combinatorics.remove.parameter}}
}
