% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.remove.R
\docType{methods}
\name{combinatorics.remove.logicconnection}
\alias{combinatorics.remove.logicconnection}
\alias{combinatorics.remove.logicconnection,Combinatorics,character-method}
\alias{combinatorics.remove.logicconnection,Combinatorics,numeric-method}
\title{Remove logicconnection}
\usage{
combinatorics.remove.logicconnection(model, logicconnection)

\S4method{combinatorics.remove.logicconnection}{Combinatorics,character}(model,
  logicconnection)

\S4method{combinatorics.remove.logicconnection}{Combinatorics,numeric}(model,
  logicconnection)
}
\arguments{
\item{model}{объект класса \code{"Combinatorics"}, содержащий комбинаторную модель}

\item{logicconnection}{объект класса \code{"character"} или \code{"numeric"}, содержащий
логическую формулу или формулы в логической сети}
}
\value{
возвращает объект класса \code{"Combinatorics"} с внесёнными изменениями
}
\description{
Исключает логическую связь из логической сети комбинаторной модели
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,logicconnection = character}: исключает логическую связь по её логической
формуле

\item \code{model = Combinatorics,logicconnection = numeric}: исключает логическую связь по индексу её логической
формулы в логической сети (в списке логических связей)
}}

\examples{

  model <- combinatorics.example.Model3()
  combinatorics.remove.logicconnection(model, "[1,4,5]")

  model <- combinatorics.example.Model3()
  combinatorics.remove.logicconnection(model, 4)
}
