% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.remove.R
\docType{methods}
\name{combinatorics.remove.parameter}
\alias{combinatorics.remove.parameter}
\alias{combinatorics.remove.parameter,Combinatorics,character-method}
\alias{combinatorics.remove.parameter,Combinatorics,numeric-method}
\title{Remove parameter}
\usage{
combinatorics.remove.parameter(model, parameter)

\S4method{combinatorics.remove.parameter}{Combinatorics,character}(model,
  parameter)

\S4method{combinatorics.remove.parameter}{Combinatorics,numeric}(model,
  parameter)
}
\arguments{
\item{model}{объект класса \code{"Combinatorics"}, содержащий комбинаторную модель}

\item{parameter}{объект класса \code{"character"} или \code{"numeric"}, содержащий
наименование показателя или номер показателя в списке показателей комбинаторной модели}
}
\value{
возвращает объект класса \code{"Combinatorics"} с внесёнными изменениями
}
\description{
Исключает показатель в мероприятиях из комбинаторной модели. При исключении показатель
  исключается из каждого мероприятия.
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,parameter = character}: исключает показатель по его наименованию

\item \code{model = Combinatorics,parameter = numeric}: исключает показатель по индексу в списке
показателей комбинаторной модели
}}

\examples{

  model <- combinatorics.example.Model3()
  combinatorics.remove.parameter(model, "R1")

  model <- combinatorics.example.Model3()
  combinatorics.remove.parameter(model, 4)
}
