% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.set.R
\docType{methods}
\name{combinatorics.set.arrangement}
\alias{combinatorics.set.arrangement}
\alias{combinatorics.set.arrangement,Combinatorics,character,vector-method}
\alias{combinatorics.set.arrangement,Combinatorics,numeric,vector-method}
\title{Set arrangement values}
\usage{
combinatorics.set.arrangement(model, arrangement, values)

\S4method{combinatorics.set.arrangement}{Combinatorics,character,vector}(model,
  arrangement, values)

\S4method{combinatorics.set.arrangement}{Combinatorics,numeric,vector}(model,
  arrangement, values)
}
\arguments{
\item{model}{Combinatorics model}

\item{arrangement}{arrangement in model}

\item{values}{new arrangement values}
}
\value{
Combinatroics model
}
\description{
Set arrangement values
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,arrangement = character,values = vector}: find arrangement by name and set new values

\item \code{model = Combinatorics,arrangement = numeric,values = vector}: find arrangement by index and set new values
}}

\examples{

  model <- combinatorics.example.Model3()
  combinatorics.set.arrangement(model, "M0", c(10, 20, 30, 40, 50, 60.6))

  model <- combinatorics.example.Model3()
  combinatorics.set.arrangement(model, 1, c(10.1, 20.2, 30.3, 40.4, 50.5, 60.6))
}
