% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.set.R
\docType{methods}
\name{combinatorics.set.limit}
\alias{combinatorics.set.limit}
\alias{combinatorics.set.limit,Combinatorics,character,numeric-method}
\alias{combinatorics.set.limit,Combinatorics,numeric,numeric-method}
\title{Set limit in model}
\usage{
combinatorics.set.limit(model, parameter, value)

\S4method{combinatorics.set.limit}{Combinatorics,character,numeric}(model,
  parameter, value)

\S4method{combinatorics.set.limit}{Combinatorics,numeric,numeric}(model,
  parameter, value)
}
\arguments{
\item{model}{Combinatorics model}

\item{parameter}{parameter in model}

\item{value}{new parameter value}
}
\value{
Combinatroics model
}
\description{
Set limit in model
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,parameter = character,value = numeric}: find resource by name and set limit value

\item \code{model = Combinatorics,parameter = numeric,value = numeric}: find resource by index and set limit value
}}

\examples{

  model <- combinatorics.example.Model3()
  combinatorics.set.limit (model, 'R1', 100)

  model <- combinatorics.example.Model3()
  combinatorics.set.limit (model, 4, 100)
}
