% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.set.R
\docType{methods}
\name{combinatorics.set.name}
\alias{combinatorics.set.name}
\alias{combinatorics.set.name,Combinatorics,character-method}
\title{Method set model name}
\usage{
combinatorics.set.name(model, name)

\S4method{combinatorics.set.name}{Combinatorics,character}(model, name)
}
\arguments{
\item{model}{объект класса \code{"Combinatorics"}, содержащий комбинаторную модель}

\item{name}{объект класса \code{"character"}, содержащий название модели}
}
\value{
возвращает объект класса \code{"Combinatorics"} с изменённым названием модели
}
\description{
Присваивает название комбинаторной модели
}
\examples{

  model <- combinatorics.example.Model3()
  model <- combinatorics.set.name (model, 'Demo Name')
  model@name
}
