% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.design.set.R
\docType{methods}
\name{combinatorics.set.target}
\alias{combinatorics.set.target}
\alias{combinatorics.set.target,Combinatorics,character,logical-method}
\alias{combinatorics.set.target,Combinatorics,numeric,logical-method}
\title{Set target in model}
\usage{
combinatorics.set.target(model, parameter, value)

\S4method{combinatorics.set.target}{Combinatorics,character,logical}(model,
  parameter, value)

\S4method{combinatorics.set.target}{Combinatorics,numeric,logical}(model,
  parameter, value)
}
\arguments{
\item{model}{Combinatorics model}

\item{parameter}{parameter in model}

\item{value}{has parameter is targer or not}
}
\value{
Combinatroics model
}
\description{
Set target in model
}
\section{Methods (by class)}{
\itemize{
\item \code{model = Combinatorics,parameter = character,value = logical}: find resource by name and set flag

\item \code{model = Combinatorics,parameter = numeric,value = logical}: find resource by index and set flag
}}

\examples{

  model <- combinatorics.example.Model3()
  combinatorics.set.target (model, 'R1', TRUE)

  model <- combinatorics.example.Model3()
  model <- combinatorics.set.target (model, 6, FALSE)
  combinatorics.set.target (model, 4, TRUE)
}
